requirements.txt
anthropic
streamlit
pandas
plotly
chromadb
fredapi
python-dotenv
requests
pyyaml
